sam - portfolio orm refactor
