+++ config/config.json
@@ -2,7 +2,7 @@
   "development": {
     "username": "root",
     "password": "",
-    "database": "test_db",
+    "database": "stock_db",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
