+++ config/connection.js
@@ -1,26 +1,26 @@
-var mysql = require("mysql");
+const mysql = require("mysql");
 
-var connection;
+let connection;
 
 if (process.env.JAWSDB_URL) {
-connection = mysql.createConnection(process.env.JAWSDB_URL);
+   connection = mysql.createConnection(process.env.JAWSDB_URL);
 }else {
-  connection = mysql.createConnection({
-  host: "localhost",
-  port: 3306,
-  user: "root",
-  password: "Charlie27!",
-  database: "stocks_db"
-});
+   connection = mysql.createConnection({
+      host: "localhost",
+      port: 3306,
+      user: "root",
+      password: "",
+      database: "stock_db"
+   });
 
-};
+}
 
-connection.connect(function(err) {
-  if (err) {
-    console.error("error connecting: " + err.stack);
-    return;
-  }
-  console.log("connected as id " + connection.threadId);
+connection.connect((err) => {
+   if (err) {
+      console.error("error connecting: " + err.stack);
+      return;
+   }
+   console.log("connected as id " + connection.threadId);
 });
 
 module.exports = connection;
